+++ store/store.go
@@ -8,6 +8,7 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"errors"
+	"expvar"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -41,6 +42,22 @@ const (
 	leaderWaitDelay     = 100 * time.Millisecond
 	appliedWaitDelay    = 100 * time.Millisecond
 )
+
+const (
+	numSnaphots = "num_snapshots"
+	numBackups  = "num_backups"
+	numRestores = "num_restores"
+)
+
+// stats captures stats for the Store.
+var stats *expvar.Map
+
+func init() {
+	stats = expvar.NewMap("store")
+	stats.Add(numSnaphots, 0)
+	stats.Add(numBackups, 0)
+	stats.Add(numRestores, 0)
+}
 
 // QueryRequest represents a query that returns rows, and does not modify
 // the database.
@@ -188,6 +205,7 @@ type Store struct {
 
 	ShutdownOnRemove  bool
 	SnapshotThreshold uint64
+	SnapshotInterval  time.Duration
 	HeartbeatTimeout  time.Duration
 	ElectionTimeout   time.Duration
 	ApplyTimeout      time.Duration
@@ -519,6 +537,7 @@ func (s *Store) Backup(leader bool) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
+	stats.Add(numBackups, 1)
 	return b, nil
 }
 
@@ -646,6 +665,9 @@ func (s *Store) raftConfig() *raft.Config {
 	if s.SnapshotThreshold != 0 {
 		config.SnapshotThreshold = s.SnapshotThreshold
 	}
+	if s.SnapshotInterval != 0 {
+		config.SnapshotInterval = s.SnapshotInterval
+	}
 	if s.HeartbeatTimeout != 0 {
 		config.HeartbeatTimeout = s.HeartbeatTimeout
 	}
@@ -758,6 +780,7 @@ func (s *Store) Snapshot() (raft.FSMSnapshot, error) {
 		return nil, err
 	}
 
+	stats.Add(numSnaphots, 1)
 	return fsm, nil
 }
 
@@ -819,11 +842,16 @@ func (s *Store) Restore(rc io.ReadCloser) error {
 		return err
 	}
 
-	return func() error {
+	err = func() error {
 		s.metaMu.Lock()
 		defer s.metaMu.Unlock()
 		return json.Unmarshal(b, &s.meta)
 	}()
+	if err != nil {
+		return err
+	}
+	stats.Add(numRestores, 1)
+	return nil
 }
 
 // RegisterObserver registers an observer of Raft events